move atomic_i_increment test to atomic_upgrade.rs
